+++ header/constants.h
@@ -1,7 +1,7 @@
 #pragma once
 
-const float PI = 3.14159265;
-const float g = 200;
+const double PI = 3.14159265;
+const double g = 200;
 
 const int WIDTH = 800;
 const int HEIGHT = 800;
